ref #2730