@afinch7 2019-08-17
I've have `deno_core_http_bench` working, but performance is awful currently.
```
Running 10s test @ http://127.0.0.1:4544/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.41ms  238.05us  11.67ms   93.35%
    Req/Sec     3.56k   277.97     3.91k    90.50%
  70784 requests in 10.01s, 3.44MB read
Requests/sec:   7074.11
Transfer/sec:    352.32KB
```
I don't even need to run a test on master to tell you that this is worse than 60% performance loss.