@afinch7 2019-08-17
Got it sorted. Calling the op id accessor for every dispatch was slowing things down a the most, but that atomic also accounted for about 15%.
This branch now
```
Running 10s test @ http://127.0.0.1:4544/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   286.08us   76.34us   2.36ms   93.77%
    Req/Sec    17.35k     1.61k   20.74k    64.85%
  348752 requests in 10.10s, 16.96MB read
Requests/sec:  34529.76
Transfer/sec:      1.68MB
```
Master
```
Running 10s test @ http://127.0.0.1:4544/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   286.14us   77.99us   2.46ms   94.04%
    Req/Sec    17.41k     1.61k   21.46k    71.78%
  349712 requests in 10.10s, 17.01MB read
Requests/sec:  34624.60
Transfer/sec:      1.68MB
```
Mostly within margin of error as far as I can tell. Maybe slightly in favor of master.