@afinch7 2019-08-17
Now that performance is sorted this is coming along well. Here are just a few nice features this will add(mainly for embedders):

* Dispatcher registries can be shared between isolates. This should make it trivial to implement multithreaded work stealing for even faster web servers.
* Dispatcher registries can be shared with the multiple dispatchers and the isolate at the same time, so we can register new dispatchers during dispatch easily.
* The dispatcher registry for a isolate can be swapped with a `&mut` reference to the isolate. This won't propagate to js code as easily yet. The information from the accessor `Deno.core.ids` will always be up to date, but checking this for every op call is very costly(see above). 
* Dispatcher instances can be shared between dispatcher registries.

I'm working on a better way to notify js of the current op id for any give op. Maybe something like this?
```
let listenOpId;
Deno.core.ids.builtins.OpListen = (id: number | undefined) => {
  listenOpId = id;
} 

/** Listens on 0.0.0.0:4500, returns rid. */
function listen() {
  if (!listenOpId) {
    throw new Error("Op not registered yet");
  }
  return sendSync(listenOpId, -1);
}
```
This should also allow async ops to `await` the op id they require with some resolvables.

I'm going to be focusing on getting cli working for now though. 

Also I'm not sure if you care much, but I had to drop the whole `trait NamedOpDispatcher: OpDispatcher` thing. Turns out rust won't let you implement a foreign trait on a generic(`impl<T: LocalDispatcherTrait> OpDispatcher for T`).