@ry 2019-08-17
This is really a mouthful: Deno.core.ids.builtins.OpListen

Can this be shortened to Deno.ops.listen ?