@afinch7 2019-08-18
> This is really a mouthful: Deno.core.ids.builtins.OpListen
> 
> Can this be shortened to Deno.ops.listen ?

I still want to include top level namespaceing(`builtins`), and `OpListen` can be changed in the op implementation(and the namespace `builtins` is also configurable).

I did remove the `core` prefix, and rename `ids` to `ops`, so the new layout looks like this: `Deno.ops.namespace.name`.
Also it now supports namespace aliasing I.E.
```
const ops = Deno.ops.someReallyLongNamespace;

let listenOpId;
ops.OpListen = id => {
  listenOpId = id;
}

/** Listens on 0.0.0.0:4500, returns rid. */
function listen() {
  return sendSync(listenOpId, -1);
}
```
And it still supports directly reading op ids I.E.
```
const ops = Deno.ops.someReallyLongNamespace;

let listenOpId;

/** Listens on 0.0.0.0:4500, returns rid. */
function listen() {
  if (listenOpId === undefined) {
    listenOpId = ops.OpListen;
  }
  return sendSync(listenOpId, -1);
}
```
This will still cost some time to evaluate the if, so it's not ideal this feature is mainly designed for debug/testing.