@bartlomieju 2019-08-18
@afinch7 why is this bit needed:
```
  if (listenOpId === undefined) {
    listenOpId = ops.OpListen;
  }
```?

Can't it be accessed directly via `Deno.ops.namespace.name`?