@afinch7 2019-08-18
> @afinch7 why is this bit needed:
> 
> ```
>   if (listenOpId === undefined) {
>     listenOpId = ops.OpListen;
>   }
> ```?
> 
> Can't it be accessed directly via `Deno.ops.namespace.name`?
> ```

You can, but if you care about performance at all you really shouldn't. This is implemented as a `get` accessor, so it's basically equivalent to calling a extra function for every dispatch(a pretty expensive one at that).