@bartlomieju 2019-08-18
> > @afinch7 why is this bit needed:
> > ```
> >   if (listenOpId === undefined) {
> >     listenOpId = ops.OpListen;
> >   }
> > ```?
> > 
> > Can't it be accessed directly via `Deno.ops.namespace.name`?
> > ```
> 
> You can, but if you care about performance at all you really shouldn't. This is implemented as a `get` accessor, so it's basically equivalent to calling a extra function for every dispatch(a pretty expensive one at that). It's slow for the same reason that `for (let i = 0; i < something.length; i++) {}` is slower than `const length = something.length; for (let i = 0; i < length; i++) {}`

I always thought that V8 can optimize away such loops... Is there a big difference in performance using first approach?