@afinch7 2019-08-18
`i < something.length` is a conditional, and in order to evaluate it we need to know the value of `something.length` which is some form of accessor. Since we evaluate `i < something.length` again each cycle(in case the length of `something` changed), this means we are calling the accessor for `something.length` every cycle. If we set `const length = something.length` we don't have to call the accessor every cycle anymore.