@bartlomieju 2019-08-18
> In our case for op ids it's a huge difference like 1/4-1/5 the speed if not slower. If you look at the benchmarks above the first one at the top is send(Deno.ops.builtins.OpListen, args, buf) and the second one is:

To be honest that sounds really strange to have 20% perf hit on object lookup. I wonder if that might be related to #2758. I still need to drill down a bit more into this PR.