@afinch7 2019-08-19
Don't merge this yet. I think I reintroduced the bug from #2776. I will see if I can get that one figured out tomorrow. It seems to only occur on linux during `deno bundle https://deno.land/std/http/file_server.ts`.