@afinch7 2019-08-19
Benchmarks from the last passing build
```
"req_per_sec": {
    "deno_tcp": 49632, 
    "deno_core_single": 56111, 
    "node_proxy_tcp": 20290, 
    "deno_proxy_tcp": 18218, 
    "deno_core_multi": 58596, 
    "deno_tcp_current_thread": 46730, 
    "hyper": 56331, 
    "node_tcp": 57969, 
    "deno_http": 18762, 
    "node_http": 25049, 
    "deno_proxy": 1670, 
    "node_proxy": 2709
  }, 
  "max_latency": {
    "deno_tcp": 8.94, 
    "deno_core_single": 0.582, 
    "node_proxy_tcp": 1.19, 
    "deno_proxy_tcp": 5.49, 
    "deno_core_multi": 1.13, 
    "deno_tcp_current_thread": 9.05, 
    "hyper": 0.421, 
    "node_tcp": 0.743, 
    "deno_http": 2.78, 
    "node_http": 1.94, 
    "deno_proxy": 28.09, 
    "node_proxy": 9.73
  }, 
```