@ry 2019-08-19
>  I think I reintroduced the bug from #2776.

My theory is that the bug is present in master, we just see it rarely. Appveyor builds fail probably 50% of the time - I think with a segfault - it may be this bug.