@afinch7 2019-08-19
I just ran `./target/release/deno bundle https://deno.land/std/http/file_server.ts` locally, and it created a segfault on the first run. This is clearly a bug, but I'm not sure if it applies to master.