@afinch7 2019-08-19
This bug doesn't apply to master, but that means that we know more specifically it's something I included in this PR.